+++ src/main/java/org/relunluck/engine/graph/UniformsMap.java
@@ -1,9 +1,10 @@
 package org.relunluck.engine.graph;
 
-import org.joml.Matrix4f;
+import org.joml.Matrix4d;
 import org.lwjgl.system.MemoryStack;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.Map;
 
 import static org.lwjgl.opengl.GL20.*;
 
@@ -38,7 +39,7 @@ public class UniformsMap {
         glUniform1i(getUniformLocation(uniformName), value);
     }
 
-    public void setUniform(String uniformName, Matrix4f value) {
+    public void setUniform(String uniformName, Matrix4d value) {
         try (MemoryStack stack = MemoryStack.stackPush()) {
             glUniformMatrix4fv(getUniformLocation(uniformName), false, value.get(stack.mallocFloat(16)));
         }
